fix(theme-classic) extract HomeBreadcrumbItem + fix swizzle bugs (#8445)
