+++ packages/docusaurus/src/commands/swizzle/actions.ts
@@ -56,7 +56,7 @@ export async function eject({
   const isDirectory = await isDir(fromPath);
   const globPattern = isDirectory
     ? // Do we really want to copy all components?
-      path.join(fromPath, '*')
+      path.join(fromPath, '**/*')
     : `${fromPath}.*`;
 
   const globPatternPosix = posixPath(globPattern);
@@ -67,6 +67,7 @@ export async function eject({
       // When ejecting JS components, we want to avoid emitting TS files
       // In particular the .d.ts files that theme build output contains
       typescript ? null : '**/*.{d.ts,ts,tsx}',
+      '**/{__fixtures__,__tests__}/*',
     ]),
   });
